requirements.txt
aiosqlite==0.21.0
numpy==2.2.5
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
pyqtgraph==0.13.7
python-dotenv==1.1.0
qasync==0.27.1
typing_extensions==4.13.2
xlrd==2.0.1
